# api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Job Seeker Copilot User Profile API",
    "version": "1.2.0",
    "description": "Producer-owned contract for authenticated current-user profile and versioned evidence operations."
  },
...
  "paths": {
    "/api/profiles/me": {
      "get": {
        "tags": [
          "User Profiles"
        ],
        "operationId": "getMyProfile",
        "summary": "Get the authenticated user's profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Profile found",
            "headers": {
              "ETag": {
                "description": "Quoted monotonic profile revision for optimistic concurrency",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "404": {
            "description": "Profile not found"
          }
        }
      },
      "put": {
        "tags": [
          "User Profiles"
        ],
        "operationId": "createOrUpdateMyProfile",
        "summary": "Create or replace the authenticated user's profile",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "Optional quoted profile revision. Omission is retained temporarily for legacy clients.",
            "schema": {
              "type": "string",
              "pattern": "^\"?[0-9]+\"?$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserProfile"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile saved",
            "headers": {
              "ETag": {
                "description": "Quoted resulting profile revision",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "400": {
            "description": "Profile validation failed"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "409": {
            "description": "Stale profile revision or concurrent profile write conflict"
          }
        }
      },
      "patch": {
        "tags": [
          "User Profiles"
        ],
        "operationId": "updateMyPreferences",
        "summary": "Update current profile preferences without replacing historical profile fields",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "If-Match",
            "in": "header",
            "required": false,
            "description": "Optional quoted profile revision for optimistic concurrency.",
            "schema": {
              "type": "string",
              "pattern": "^\"?[0-9]+\"?$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfilePreferencesUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Preferences saved while legacy roles and qualifications are preserved",
            "headers": {
              "ETag": {
                "description": "Quoted resulting profile revision",
                "schema": {
                  "type": "string"
                }
              }
            },
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserProfile"
                }
              }
            }
          },
          "400": {
            "description": "Preference validation failed"
          },
          "401": {
            "description": "Missing or invalid access token"
          },
          "409": {
            "description": "Stale profile revision"
          }
        }
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "UserProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "userId": {
            "type": "string",
            "maxLength": 128,
            "readOnly": true
          },
          "revision": {
            "type": "integer",
            "format": "int64",
            "minimum": 1,
            "readOnly": true,
            "description": "Monotonic revision of claimant-declared profile content"
          },
          "revisionId": {
            "type": "string",
            "format": "uuid",
            "readOnly": true,
            "description": "Immutable opaque identity for this profile revision"
          },
          "contentDigest": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$",
            "readOnly": true,
            "description": "SHA-256 digest of normalized claimant-declared profile content"
          },
          "skills": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          "aspirations": {
            "$ref": "#/components/schemas/Aspirations"
          },
          "workPreferences": {
            "$ref": "#/components/schemas/WorkPreferences"
          },
          "qualifications": {
            "type": "array",
            "maxItems": 50,
            "items": {
              "$ref": "#/components/schemas/Qualification"
            }
          },
          "roles": {
            "type": "array",
            "maxItems": 50,
            "items": {
              "$ref": "#/components/schemas/Role"
            }
          }
        }
      },
      "Aspirations": {
        "type": "object",
        "properties": {
          "targetRoles": {
            "type": "array",
            "maxItems": 50,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          "targetWeeklyHours": {
            "$ref": "#/components/schemas/TargetWeeklyHours"
          }
        }
      },
      "TargetWeeklyHours": {
        "type": "string",
        "enum": [
          "FULL_TIME",
          "PART_TIME_16_30",
          "PART_TIME_UNDER_16",
          "FLEXIBLE"
        ]
      },
      "WorkPreferences": {
        "type": "object",
        "properties": {
          "location": {
            "$ref": "#/components/schemas/PostcodeLocation"
          },
          "commuteRange": {
            "type": "integer",
            "minimum": 0,
            "maximum": 500
          },
          "employmentTypes": {
            "type": "array",
            "maxItems": 5,
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/EmploymentType"
            }
          },
          "workingPatterns": {
            "type": "array",
            "maxItems": 8,
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/WorkingPattern"
            }
          },
          "workplaceArrangements": {
            "type": "array",
            "maxItems": 3,
            "uniqueItems": true,
            "items": {
              "$ref": "#/components/schemas/WorkplaceArrangement"
            }
          },
          "availableFrom": {
            "type": "string",
            "format": "date",
            "description": "Explicit claimant-selected availability date; never fabricated"
          },
          "noticePeriodDays": {
            "type": "integer",
            "minimum": 0,
            "maximum": 3650,
            "description": "Explicit claimant-selected notice period; mutually exclusive with availableFrom"
          }
        }
      },
      "ProfilePreferencesUpdate": {
        "type": "object",
        "properties": {
          "skills": {
            "type": "array",
            "maxItems": 100,
            "items": {
              "type": "string",
              "minLength": 1,
              "maxLength": 100
            }
          },
          "aspirations": {
            "$ref": "#/components/schemas/Aspirations"
          },
          "workPreferences": {
            "$ref": "#/components/schemas/WorkPreferences"
          }
        },
        "description": "Current intentions and preferences only; roles and qualifications are preserved"
      },
      "EmploymentType": {
        "type": "string",
        "enum": [
          "PERMANENT",
          "FIXED_TERM",
          "TEMPORARY",
          "APPRENTICESHIP",
          "CONTRACT"
        ]
      },
      "WorkingPattern": {
        "type": "string",
        "enum": [
          "FULL_TIME",
          "PART_TIME",
          "FLEXIBLE",
          "DAY",
          "EVENING",
          "NIGHT",
          "WEEKEND",
          "SHIFT"
        ]
      },
      "WorkplaceArrangement": {
        "type": "string",
        "enum": [
          "ONSITE",
          "HYBRID",
          "REMOTE"
        ]
      },
      "PostcodeLocation": {
        "type": "object",
        "properties": {
          "postcode": {
            "type": "string",
            "maxLength": 16,
            "description": "UK full postcode or outcode; normalized by the service"
          },
          "region": {
            "type": "string",
            "maxLength": 100
          },
          "adminDistrict": {
            "type": "string",
            "maxLength": 100
          },
          "latitude": {
            "type": "number",
            "format": "double",
            "minimum": -90,
            "maximum": 90
          },
          "longitude": {
            "type": "number",
            "format": "double",
            "minimum": -180,
            "maximum": 180
          }
        }
      },
      "Qualification": {
        "type": "object",
        "required": [
          "qualificationName",
          "issuingBody",
          "status"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "qualificationName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "issuingBody": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "status": {
            "$ref": "#/components/schemas/QualificationStatus"
          },
          "grade": {
            "type": "string",
            "maxLength": 100
          },
          "dateAchieved": {
            "type": "string",
            "maxLength": 30,
            "description": "YYYY-MM or YYYY-MM-DD"
          },
          "expectedCompletion": {
            "type": "string",
            "maxLength": 30,
            "description": "YYYY-MM or YYYY-MM-DD"
          }
        }
      },
      "QualificationStatus": {
        "type": "string",
        "enum": [
          "IN_PROGRESS",
          "COMPLETED"
        ]
      },
      "Role": {
        "type": "object",
        "required": [
          "jobTitle",
          "employer",
          "status",
          "startDate"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64",
            "readOnly": true
          },
          "jobTitle": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "employer": {
            "type": "string",
            "minLength": 1,
            "maxLength": 200
          },
          "status": {
            "$ref": "#/components/schemas/RoleStatus"
          },
          "startDate": {
            "type": "string",
            "minLength": 1,
            "maxLength": 30,
            "description": "YYYY-MM or YYYY-MM-DD"
          },
          "endDate": {
            "type": "string",
            "maxLength": 30,
            "description": "YYYY-MM or YYYY-MM-DD"
          },
          "keyResponsibilities": {
            "type": "string",
            "maxLength": 2000
          }
        }
      },
      "RoleStatus": {
        "type": "string",
        "enum": [
          "CURRENT",
          "PREVIOUS_ROLE"
        ]
      },
...
    }
  }
}